code style and some amendments by Spring Boot Security app done

+++ src/main/java/ru/kata/spring/boot_bootstrap/demo/dao/UserDaoImpl.java
@@ -1,25 +1,21 @@
 package ru.kata.spring.boot_bootstrap.demo.dao;
 
+import org.springframework.stereotype.Repository;
+import ru.kata.spring.boot_bootstrap.demo.models.User;
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
-
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-import ru.kata.spring.boot_bootstrap.demo.models.User;
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 
 @Repository
 public class UserDaoImpl implements UserDao {
-
    @PersistenceContext
    private EntityManager entityManager;
 
    @Override
-   @Transactional(readOnly = true)
    public User findUserByEmail(String email) {
       TypedQuery<User> query = entityManager.createQuery(
               "SELECT u FROM User u LEFT JOIN FETCH u.roles WHERE u.email =: email", User.class); // запрос JPQL
@@ -32,32 +28,36 @@ public class UserDaoImpl implements UserDao {
    }
 
    @Override
-   @Transactional(readOnly = true)
    public List<User> getUsersList() {
       return entityManager.createQuery("SELECT u FROM User u", User.class).getResultList();
    }
 
    @Override
-   @Transactional(readOnly = true)
    public User getUserById(Long id) {
+      if (entityManager.find(User.class, id) == null) {
+         throw new EntityNotFoundException(String.format("User with ID '%s'  not found", id));
+      }
       return entityManager.find(User.class, id);
    }
 
    @Override
-   @Transactional
    public void addUser(User user) {
       entityManager.persist(user);
    }
 
    @Override
-   @Transactional
    public void updateUser(User user) {
+      if (entityManager.find(User.class, user.getId()) == null) {
+         throw new EntityNotFoundException(String.format("User with ID '%s'  not found", user.getId()));
+      }
       entityManager.merge(user);
    }
 
    @Override
-   @Transactional
    public void deleteUser(User user) {
+      if (entityManager.find(User.class, user.getId()) == null) {
+         throw new EntityNotFoundException(String.format("User with ID '%s'  not found", user.getId()));
+      }
       entityManager.remove(user);
    }
 }
